+++ lib/oauth2_client.dart
@@ -3,11 +3,20 @@ import 'dart:convert';
 import 'package:http/http.dart' as http;
 import 'package:oauth2_client/access_token_response.dart';
 import 'package:oauth2_client/authorization_response.dart';
-import 'package:meta/meta.dart';
 import 'package:oauth2_client/oauth2_response.dart';
 import 'package:oauth2_client/src/oauth2_utils.dart';
-import 'package:oauth2_client/src/web_auth.dart';
+// import 'package:oauth2_client/src/web_auth.dart';
+
+import 'src/base_web_auth.dart';
+import 'src/web_auth.dart'
+    // ignore: uri_does_not_exist
+    if (dart.library.io) 'src/io_web_auth.dart'
+    // ignore: uri_does_not_exist
+    if (dart.library.html) 'src/browser_web_auth.dart';
+
 import 'package:random_string/random_string.dart';
+
+enum CredentialsLocation { HEADER, BODY }
 
 /// Base class that implements OAuth2 authorization flows.
 ///
@@ -34,32 +43,44 @@ class OAuth2Client {
   String customUriScheme;
 
   String tokenUrl;
-  String refreshUrl;
-  String revokeUrl;
+  String? refreshUrl;
+  String? revokeUrl;
   String authorizeUrl;
+  String scopeSeparator;
 
-  Map<String, String> _accessTokenRequestHeaders;
+  Map<String, String> _accessTokenRequestHeaders = {};
 
-  WebAuth webAuthClient;
+  BaseWebAuth webAuthClient = createWebAuth();
+  CredentialsLocation credentialsLocation;
 
+  /// Creates a new client instance with the following parameters:
+  ///
+  /// * [authorizeUrl]: the url that must be used to fetch authorization codes (for Authorization Code flows)
+  /// * [tokenUrl]: the url to be used for generating the OAuth2 access tokens
+  /// * [refreshUrl]: the url that must be used for refreshing an Access Token
+  /// * [revokeUrl]: the url to be invoked for token revocation
+  /// * [redirectUri]: the redirect uri defined in the provider's client registration panel
+  /// * [customUriScheme]: the scheme used for the redirect uri
+  /// * [credentialsLocation]: where the credentials (client ID / client secret) should be passed (header / body)
+  /// * [scopeSeparator]: the separator that has to be used to serialize scopes in the token request
   OAuth2Client(
-      {@required this.authorizeUrl,
-      @required this.tokenUrl,
+      {required this.authorizeUrl,
+      required this.tokenUrl,
       this.refreshUrl,
       this.revokeUrl,
-      @required this.redirectUri,
-      @required this.customUriScheme}) {
-    webAuthClient = WebAuth();
-  }
+      required this.redirectUri,
+      required this.customUriScheme,
+      this.credentialsLocation = CredentialsLocation.HEADER,
+      this.scopeSeparator = '+'});
 
   /// Requests an Access Token to the OAuth2 endpoint using the Implicit grant flow (https://tools.ietf.org/html/rfc6749#page-31)
   Future<AccessTokenResponse> getTokenWithImplicitGrantFlow({
-    @required String clientId,
-    List<String> scopes,
+    required String clientId,
+    List<String>? scopes,
     bool enableState = true,
-    String state,
+    String? state,
     httpClient,
-    webAuthClient,
+    BaseWebAuth? webAuthClient,
   }) async {
     httpClient ??= http.Client();
     webAuthClient ??= this.webAuthClient;
@@ -76,7 +97,9 @@ class OAuth2Client {
 
     // Present the dialog to the user
     final result = await webAuthClient.authenticate(
-        url: authorizeUrl, callbackUrlScheme: customUriScheme);
+        url: authorizeUrl,
+        callbackUrlScheme: customUriScheme,
+        redirectUrl: redirectUri);
 
     final fragment = Uri.splitQueryString(Uri.parse(result).fragment);
 
@@ -99,22 +122,22 @@ class OAuth2Client {
 
   /// Requests an Access Token to the OAuth2 endpoint using the Authorization Code Flow.
   Future<AccessTokenResponse> getTokenWithAuthCodeFlow({
-    @required String clientId,
-    List<String> scopes,
-    String clientSecret,
+    required String clientId,
+    List<String>? scopes,
+    String? clientSecret,
     bool enablePKCE = true,
     bool enableState = true,
-    String state,
-    String codeVerifier,
-    Function afterAuthorizationCodeCb,
-    Map<String, dynamic> authCodeParams,
-    Map<String, dynamic> accessTokenParams,
+    String? state,
+    String? codeVerifier,
+    Function? afterAuthorizationCodeCb,
+    Map<String, dynamic>? authCodeParams,
+    Map<String, dynamic>? accessTokenParams,
     httpClient,
-    webAuthClient,
+    BaseWebAuth? webAuthClient,
   }) async {
-    AccessTokenResponse tknResp;
+    AccessTokenResponse? tknResp;
 
-    String codeChallenge;
+    String? codeChallenge;
 
     if (enablePKCE) {
       codeVerifier ??= randomAlphaNumeric(80);
@@ -136,12 +159,16 @@ class OAuth2Client {
 
       tknResp = await requestAccessToken(
           httpClient: httpClient,
-          code: authResp.code,
+          //If the authorization request was successfull, the code must be set
+          //otherwise an exception is raised in the OAuth2Response constructor
+          code: authResp.code!,
           clientId: clientId,
           scopes: scopes,
           clientSecret: clientSecret,
           codeVerifier: codeVerifier,
           customParams: accessTokenParams);
+    } else {
+      tknResp = AccessTokenResponse.errorResponse();
     }
 
     return tknResp;
@@ -149,13 +176,15 @@ class OAuth2Client {
 
   /// Requests an Access Token to the OAuth2 endpoint using the Client Credentials flow.
   Future<AccessTokenResponse> getTokenWithClientCredentialsFlow(
-      {@required String clientId,
-      @required String clientSecret,
-      List<String> scopes,
+      {required String clientId,
+      required String clientSecret,
+      List<String>? scopes,
       httpClient}) async {
     var params = <String, String>{'grant_type': 'client_credentials'};
 
-    if (scopes != null) params['scope'] = scopes.map((s) => s.trim()).join('+');
+    if (scopes != null && scopes.isNotEmpty) {
+      params['scope'] = serializeScopes(scopes);
+    }
 
     var response = await _performAuthorizedRequest(
         url: tokenUrl,
@@ -164,19 +193,18 @@ class OAuth2Client {
         params: params,
         httpClient: httpClient);
 
-    return AccessTokenResponse.fromHttpResponse(response,
-        requestedScopes: scopes);
+    return http2TokenResponse(response, requestedScopes: scopes);
   }
 
   /// Requests an Authorization Code to be used in the Authorization Code grant.
   Future<AuthorizationResponse> requestAuthorization({
-    @required String clientId,
-    List<String> scopes,
-    String codeChallenge,
+    required String clientId,
+    List<String>? scopes,
+    String? codeChallenge,
     bool enableState = true,
-    String state,
-    Map<String, dynamic> customParams,
-    webAuthClient,
+    String? state,
+    Map<String, dynamic>? customParams,
+    BaseWebAuth? webAuthClient,
   }) async {
     webAuthClient ??= this.webAuthClient;
 
@@ -195,19 +223,21 @@ class OAuth2Client {
 
     // Present the dialog to the user
     final result = await webAuthClient.authenticate(
-        url: authorizeUrl, callbackUrlScheme: customUriScheme);
+        url: authorizeUrl,
+        callbackUrlScheme: customUriScheme,
+        redirectUrl: redirectUri);
 
     return AuthorizationResponse.fromRedirectUri(result, state);
   }
 
   /// Requests and Access Token using the provided Authorization [code].
   Future<AccessTokenResponse> requestAccessToken(
-      {@required String code,
-      @required String clientId,
-      String clientSecret,
-      String codeVerifier,
-      List<String> scopes,
-      Map<String, dynamic> customParams,
+      {required String code,
+      required String clientId,
+      String? clientSecret,
+      String? codeVerifier,
+      List<String>? scopes,
+      Map<String, dynamic>? customParams,
       httpClient}) async {
     final params = getTokenUrlParams(
         code: code,
@@ -223,17 +253,12 @@ class OAuth2Client {
         headers: _accessTokenRequestHeaders,
         httpClient: httpClient);
 
-    return constructAccessTokenResponse(response, requestedScopes: scopes);
-  }
-
-  AccessTokenResponse constructAccessTokenResponse(http.Response response, {requestedScopes}) {
-    return AccessTokenResponse.fromHttpResponse(response,
-        requestedScopes: requestedScopes);
+    return http2TokenResponse(response, requestedScopes: scopes);
   }
 
   /// Refreshes an Access Token issuing a refresh_token grant to the OAuth2 server.
   Future<AccessTokenResponse> refreshToken(String refreshToken,
-      {httpClient, String clientId, String clientSecret}) async {
+      {httpClient, required String clientId, String? clientSecret}) async {
     final Map params = getRefreshUrlParams(refreshToken: refreshToken);
 
     var response = await _performAuthorizedRequest(
@@ -243,12 +268,12 @@ class OAuth2Client {
         params: params,
         httpClient: httpClient);
 
-    return constructAccessTokenResponse(response);
+    return http2TokenResponse(response);
   }
 
   /// Revokes both the Access and the Refresh tokens in the provided [tknResp]
   Future<OAuth2Response> revokeToken(AccessTokenResponse tknResp,
-      {String clientId, String clientSecret, httpClient}) async {
+      {String? clientId, String? clientSecret, httpClient}) async {
     var tokenRevocationResp = await revokeAccessToken(tknResp,
         clientId: clientId, clientSecret: clientSecret, httpClient: httpClient);
     if (tokenRevocationResp.isValid()) {
@@ -263,28 +288,28 @@ class OAuth2Client {
 
   /// Revokes the Access Token in the provided [tknResp]
   Future<OAuth2Response> revokeAccessToken(AccessTokenResponse tknResp,
-      {String clientId, String clientSecret, httpClient}) async {
+      {String? clientId, String? clientSecret, httpClient}) async {
     return await _revokeTokenByType(tknResp, 'access_token',
         clientId: clientId, clientSecret: clientSecret, httpClient: httpClient);
   }
 
   /// Revokes the Refresh Token in the provided [tknResp]
   Future<OAuth2Response> revokeRefreshToken(AccessTokenResponse tknResp,
-      {String clientId, String clientSecret, httpClient}) async {
+      {String? clientId, String? clientSecret, httpClient}) async {
     return await _revokeTokenByType(tknResp, 'refresh_token',
         clientId: clientId, clientSecret: clientSecret, httpClient: httpClient);
   }
 
   /// Generates the url to be used for fetching the authorization code.
   String getAuthorizeUrl(
-      {@required String clientId,
+      {required String clientId,
       String responseType = 'code',
-      String redirectUri,
-      List<String> scopes,
+      String? redirectUri,
+      List<String>? scopes,
       bool enableState = true,
-      String state,
-      String codeChallenge,
-      Map<String, dynamic> customParams}) {
+      String? state,
+      String? codeChallenge,
+      Map<String, dynamic>? customParams}) {
     final params = <String, dynamic>{
       'response_type': responseType,
       'client_id': clientId
@@ -294,7 +319,9 @@ class OAuth2Client {
       params['redirect_uri'] = redirectUri;
     }
 
-    handleScopes(params, scopes);
+    if (scopes != null && scopes.isNotEmpty) {
+      params['scope'] = serializeScopes(scopes);
+    }
 
     if (enableState && state != null && state.isNotEmpty) {
       params['state'] = state;
@@ -312,16 +339,12 @@ class OAuth2Client {
     return OAuth2Utils.addParamsToUrl(authorizeUrl, params);
   }
 
-  void handleScopes(Map<String, dynamic> params, List<String> scopes) {
-    if (scopes != null && scopes.isNotEmpty) params['scope'] = scopes;
-  }
-
   /// Returns the parameters needed for the authorization code request
   Map<String, dynamic> getTokenUrlParams(
-      {@required String code,
-      String redirectUri,
-      String codeVerifier,
-      Map<String, dynamic> customParams}) {
+      {required String code,
+      String? redirectUri,
+      String? codeVerifier,
+      Map<String, dynamic>? customParams}) {
     final params = <String, dynamic>{
       'grant_type': 'authorization_code',
       'code': code
@@ -354,25 +377,35 @@ class OAuth2Client {
   /// Performs a post request to the specified [url],
   /// adding authentication credentials as described here: https://tools.ietf.org/html/rfc6749#section-2.3
   Future<http.Response> _performAuthorizedRequest(
-      {@required String url,
-      @required String clientId,
-      String clientSecret,
-      Map params,
-      Map<String, String> headers,
+      {required String url,
+      required String clientId,
+      String? clientSecret,
+      Map? params,
+      Map<String, String>? headers,
       httpClient}) async {
     httpClient ??= http.Client();
 
     headers ??= {};
+    params ??= {};
 
     //If a client secret has been specified, it will be sent in the "Authorization" header instead of a body parameter...
-    if (clientSecret == null || clientSecret.isEmpty) {
-      if (clientId != null && clientId.isNotEmpty) {
+    if (clientSecret == null) {
+      if (clientId.isNotEmpty) {
         params['client_id'] = clientId;
       }
     } else {
-      var authHeaders = getAuthorizationHeader(
-          clientId: clientId, clientSecret: clientSecret);
-      headers.addAll(authHeaders);
+      switch (credentialsLocation) {
+        case CredentialsLocation.HEADER:
+          headers.addAll(getAuthorizationHeader(
+            clientId: clientId,
+            clientSecret: clientSecret,
+          ));
+          break;
+        case CredentialsLocation.BODY:
+          params['client_id'] = clientId;
+          params['client_secret'] = clientSecret;
+          break;
+      }
     }
 
     var response =
@@ -382,11 +415,10 @@ class OAuth2Client {
   }
 
   Map<String, String> getAuthorizationHeader(
-      {String clientId, String clientSecret}) {
+      {required String clientId, String? clientSecret}) {
     var headers = <String, String>{};
 
-    if ((clientId != null && clientId.isNotEmpty) &&
-        (clientSecret != null && clientSecret.isNotEmpty)) {
+    if ((clientId.isNotEmpty) && (clientSecret != null)) {
       var credentials =
           base64.encode(utf8.encode(clientId + ':' + clientSecret));
 
@@ -397,35 +429,29 @@ class OAuth2Client {
   }
 
   /// Returns the parameters needed for the refresh token request
-  Map<String, String> getRefreshUrlParams(
-      {@required String refreshToken, String clientId, String clientSecret}) {
+  Map<String, String> getRefreshUrlParams({required String refreshToken}) {
     final params = <String, String>{
       'grant_type': 'refresh_token',
       'refresh_token': refreshToken
     };
 
-/*
-    if (clientSecret == null || clientSecret.isEmpty) {
-      if (clientId != null && clientId.isNotEmpty) {
-        params['client_id'] = clientId;
-      }
-    }
-    if (clientId != null && clientId.isNotEmpty) {
-      params['client_id'] = clientId;
-    }
-
-    if (clientSecret != null && clientSecret.isNotEmpty) {
-      params['client_secret'] = clientSecret;
-    }
-*/
-
     return params;
+  }
+
+  AccessTokenResponse http2TokenResponse(http.Response response,
+      {List<String>? requestedScopes}) {
+    return AccessTokenResponse.fromHttpResponse(response,
+        requestedScopes: requestedScopes);
+  }
+
+  String serializeScopes(List<String> scopes) {
+    return scopes.map((s) => s.trim()).join(scopeSeparator);
   }
 
   /// Revokes the specified token [type] in the [tknResp]
   Future<OAuth2Response> _revokeTokenByType(
       AccessTokenResponse tknResp, String tokenType,
-      {String clientId, String clientSecret, httpClient}) async {
+      {String? clientId, String? clientSecret, httpClient}) async {
     var resp = OAuth2Response();
 
     if (revokeUrl == null) return resp;
@@ -443,9 +469,9 @@ class OAuth2Client {
       if (clientSecret != null) params['client_secret'] = clientSecret;
 
       http.Response response =
-          await httpClient.post(Uri.parse(revokeUrl), body: params);
+          await httpClient.post(Uri.parse(revokeUrl!), body: params);
 
-      resp = OAuth2Response.fromHttpResponse(response);
+      resp = http2TokenResponse(response);
     }
 
     return resp;