+++ lib/src/secure_storage.dart
@@ -1,13 +1,15 @@
-import 'package:oauth2_client/src/storage.dart';
+import 'package:oauth2_client/src/base_storage.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
-class SecureStorage extends Storage {
+BaseStorage createStorage() => SecureStorage();
+
+class SecureStorage implements BaseStorage {
   static final FlutterSecureStorage storage = FlutterSecureStorage();
 
   SecureStorage();
 
   @override
-  Future<String> read(String key) async {
+  Future<String?> read(String key) async {
     final options = IOSOptions(accessibility: IOSAccessibility.first_unlock);
     return await storage.read(key: key, iOptions: options);
   }