+++ lib/src/token_storage.dart
@@ -1,21 +1,26 @@
 import 'dart:convert';
 import 'package:oauth2_client/access_token_response.dart';
-import 'package:oauth2_client/src/secure_storage.dart';
-import 'package:oauth2_client/src/storage.dart';
+
+import 'base_storage.dart';
+import 'storage.dart'
+    // ignore: uri_does_not_exist
+    if (dart.library.io) 'secure_storage.dart'
+    // ignore: uri_does_not_exist
+    if (dart.library.html) 'browser_storage.dart';
 
 class TokenStorage {
   String key;
 
-  Storage storage;
+  BaseStorage storage = createStorage();
 
-  TokenStorage(this.key, {this.storage}) {
-    storage ??= SecureStorage();
+  TokenStorage(this.key, {BaseStorage? storage}) {
+    if (storage != null) this.storage = storage;
   }
 
   /// Looks for a token in the storage that matches the required [scopes].
   /// If a token in the storage has been generated for a superset of the requested scopes, it is considered valid.
-  Future<AccessTokenResponse> getToken(List<String> scopes) async {
-    AccessTokenResponse tknResp;
+  Future<AccessTokenResponse?> getToken(List<String> scopes) async {
+    AccessTokenResponse? tknResp;
 
     final serializedStoredTokens = await storage.read(key);
 
@@ -28,14 +33,14 @@ class TokenStorage {
       var tknMap = storedTokens.values.firstWhere((tkn) {
         var found = false;
 
-        if (cleanScopes == null || cleanScopes.isEmpty) {
+        if (cleanScopes.isEmpty) {
           //If the scopes are empty, onlty tokens granted to empty scopes are considered valid...
           found = (tkn['scope'] == null || tkn['scope'].isEmpty);
         } else {
           //...Otherwise look for a token granted to a superset of the requested scopes
           final tknCleanScopes = clearScopes(tkn['scope'].cast<String>());
 
-          if (tknCleanScopes != null) {
+          if (tknCleanScopes.isNotEmpty) {
             var s1 = Set.from(tknCleanScopes);
             var s2 = Set.from(cleanScopes);
             found = s1.intersection(s2).length == cleanScopes.length;
@@ -62,7 +67,7 @@ class TokenStorage {
 
   Future<Map<String, Map>> insertToken(AccessTokenResponse tknResp) async {
     final serTokens = await storage.read(key);
-    final scopeKey = getScopeKey(tknResp.scope);
+    final scopeKey = getScopeKey(tknResp.scope ?? []);
     var tokens = <String, Map>{};
 
     if (serTokens != null) {
@@ -100,8 +105,9 @@ class TokenStorage {
     return true;
   }
 
-  List clearScopes(List<String> scopes) {
-    return scopes?.where((element) => element.trim().isNotEmpty)?.toList();
+  List<String> clearScopes(List<String> scopes) {
+    // return scopes?.where((element) => element.trim().isNotEmpty)?.toList();
+    return scopes.where((element) => element.trim().isNotEmpty).toList();
   }
 
   List getSortedScopes(List<String> scopes) {
@@ -109,7 +115,7 @@ class TokenStorage {
 
     var cleanScopes = clearScopes(scopes);
 
-    if (cleanScopes != null) {
+    if (cleanScopes.isNotEmpty) {
       sortedScopes = cleanScopes.toList()
         ..sort((a, b) => a.toLowerCase().compareTo(b.toLowerCase()));
     }